"feat: improve on stories duplication, building tree, traverseAndCreate function"
In this commit, support for updating stories and forcing update is introduced in the story API. The functionality differentiates between stories to be created and stories to be updated. This information is reflected in the TreeNode type with the addition of an action key. Further, the application of 'force_update' option in story update is also added for insuring the updates even when conflicts occur. This commit enhances the capability of the story API by offering more flexibility in handling story objects.